Use regex-based image URL parsing over selectors
This skips having to parse the *entire* HTML document and only searches
for URLs that match the given regex. As a result, this part of scraping
should be significantly faster.

This also removes the UTF-8 check of the entire HTML document and
restricts it to only where it's needed: the URL itself.